Move refinement_matrices to multi_grid
